move part syncer files
