+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Parts/BagPartGraphSyncer.cs
@@ -13,7 +13,7 @@ using OrchardCore.ContentManagement.Metadata;
 using OrchardCore.ContentManagement.Metadata.Models;
 using OrchardCore.Flows.Models;
 
-namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Parts
 {
     public class BagPartGraphSyncer : IContentPartGraphSyncer
     {
